refactor: project structure update part 2
