+++ server/src/test/kotlin/com/appsheaven/keepup/database/todo/infrastructure/persistence/MongoTodoDatabaseTest.kt
@@ -1,6 +1,7 @@
-package com.appsheaven.keepup.database
+package com.appsheaven.keepup.database.todo.infrastructure.persistence
 
-import com.appsheaven.keepup.todos.infrastructure.MongoTodoDatabase
+import com.appsheaven.keepup.database.todo.domain.TodoFixtures
+import com.appsheaven.keepup.todos.infrastructure.persistence.MongoTodoDatabase
 import com.mongodb.MongoException
 import com.mongodb.kotlin.client.coroutine.MongoClient
 import kotlinx.coroutines.test.runTest
@@ -14,7 +15,7 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 
-class MongoDatabaseTest {
+class MongoTodoDatabaseTest {
 
     private lateinit var dbContainer: MongoDBContainer
     private lateinit var client: MongoClient
@@ -34,7 +35,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `inserts new todo`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
 
         val actual = sut.insertTodo(todo)
 
@@ -44,7 +45,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `throws on inserting todo with error`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
 
         sut.insertTodo(todo)
 
@@ -53,7 +54,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `gets all todos`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
         val anotherTodo = todo.copy(_id = ObjectId())
 
         sut.insertTodo(todo)
@@ -73,7 +74,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `gets todo with pagination for page 1`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
         val anotherTodo = todo.copy(_id = ObjectId())
 
         sut.insertTodo(todo)
@@ -86,7 +87,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `gets todo with pagination for page 2`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
         val anotherTodo = todo.copy(_id = ObjectId())
 
         sut.insertTodo(todo)
@@ -99,7 +100,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `return empty list with single todo for page 2`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
 
         sut.insertTodo(todo)
 
@@ -110,7 +111,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `puts todo`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
 
         sut.insertTodo(todo)
 
@@ -123,7 +124,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `deletes todo by id`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
 
         sut.insertTodo(todo)
 
@@ -134,7 +135,7 @@ class MongoDatabaseTest {
 
     @Test
     fun `deletes todo by id with no match`() = runTest {
-        val todo = MongoDatabaseFixtures.anyTodo
+        val todo = TodoFixtures.anyTodo
 
         sut.insertTodo(todo)
